refactoring by moving functions into modules
