+++ options.html
@@ -105,6 +105,5 @@
       </p>
     </div>
   </body>
-  <script src="helper_js/permissions-manager.js"></script>
-  <script src="options.js"></script>
+  <script type="module" src="options.js"></script>
 </html>
